requirements.txt
requests
pandas
apache-airflow
dbt
psycopg2-binary
sqlalchemy